+++ components/nfceditor/impl/src/main/java/com/flipperdevices/nfceditor/impl/viewmodel/NfcEditorViewModel.kt
@@ -16,7 +16,9 @@ import com.flipperdevices.nfceditor.impl.model.NfcEditorState
 import com.github.terrakok.cicerone.Router
 import javax.inject.Inject
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 
 class NfcEditorViewModel(
@@ -34,6 +36,8 @@ class NfcEditorViewModel(
     lateinit var synchronizationApi: SynchronizationApi
 
     private val textUpdaterHelper = TextUpdaterHelper()
+
+    private val showOnSaveDialogState = MutableStateFlow(false)
 
     private var isDirty = false
 
@@ -58,17 +62,29 @@ class NfcEditorViewModel(
 
     fun getNfcEditorState(): StateFlow<NfcEditorState?> = textUpdaterHelper.getNfcEditorState()
 
+    fun getShowOnSaveDialogState(): StateFlow<Boolean> = showOnSaveDialogState
+
+    fun dismissDialog() = showOnSaveDialogState.update { false }
+
     fun onCellFocus(location: NfcEditorCellLocation?) {
         textUpdaterHelper.onSelectCell(location)
-    }
-
-    fun isDirty(): Boolean {
-        return isDirty
     }
 
     fun onKeyInput(hexKey: HexKey) {
         isDirty = true
         textUpdaterHelper.onKeyboardPress(hexKey)
+    }
+
+    fun onBack(router: Router) {
+        if (currentActiveCell != null) {
+            onCellFocus(null)
+            return
+        }
+        if (isDirty) {
+            showOnSaveDialogState.update { true }
+            return
+        }
+        router.exit()
     }
 
     fun onSave(router: Router) {
